requirements.txt
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
filelock==3.18.0
fsspec==2025.5.0
huggingface-hub==0.31.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
kaggle==1.7.4.5
kagglehub==0.3.12
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==6.31.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
sympy==1.14.0
text-unidecode==1.3
tifffile==2025.5.10
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
torchxrayvision==1.3.4
tqdm==4.67.1
transformers==4.52.2
triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
